+++ registry/src/endpoints/deprecate_version.rs
@@ -36,7 +36,7 @@ pub async fn deprecate_package_version(
 		String::new()
 	};
 	let name = path.into_inner();
-	let source = app_state.source.lock().await;
+	let source = app_state.source.write().await;
 
 	let Some(scope_info) = read_scope_info(&app_state, name.scope(), &source).await? else {
 		return Ok(HttpResponse::NotFound().finish());