+++ registry/CHANGELOG.md
@@ -8,9 +8,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Added
 - Support deprecating and yanking packages by @daimond113
+- Add yanking & deprecating to registry by @daimond113
+- Log more information about configured auth & storage by @daimond113
 
 ### Changed
 - Asyncify blocking operations by @daimond113
+
+### Performance
+- Switch to using a `RwLock` over a `Mutex` to store repository data by @daimond113
 
 ## [0.1.2]
 ### Changed
@@ -18,7 +23,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [0.1.1] - 2024-12-19
 ### Changed
-- Switch to traccing for logging by @daimond113
+- Switch to tracing for logging by @daimond113
 
 ## [0.1.0] - 2024-12-14
 ### Added